Decompile SEL func_801B3164 Co-authored-by: Luciano Ciccariello <Xeeynamo@users.noreply.github.com>
