+++ config/symbols.us.stsel.txt
@@ -3,20 +3,12 @@ g_Cluts = 0x8018070C;
 g_EntityGfxs = 0x80180760;
 g_StreamWidth = 0x8018B2FC;
 g_StreamHeight = 0x8018B300;
-PrintFileSelectPlaceName = 0x801AD134;
 Update = 0x801AEED8;
-SetTexturedPrimRect = 0x801B26A0;
-DrawString16x16 = 0x801B29C0;
-DrawImages8x8 = 0x801B2AFC;
 HydrateSaveSummaryEntry = 0x801B2F50;
 MemcardInit = 0x801B84F0;
-MemcardParse = 0x801B8518;
-MemcardDetectSave = 0x801B873C;
-IsMemcardBlockUsed = 0x801B881C;
 MemcardReadFile = 0x801B884C;
 MemcardWriteFile = 0x801B88F4;
 MemcardClose = 0x801B8A10;
-func_800E9880 = 0x801B8A8C;
 StoreSaveData = 0x801B8E20;
 LoadSaveData = 0x801B94BC;
 MakeMemcardPath = 0x801B9698;
